Add files via upload
Correction of test to draw object rectangle ("yes" was tested, had to be "Yes"). Label will be drawn inside the Rectangle box if there is no place anymore for the label above the left top of the box.  Recording of Stream1 works correct on a server with enough power. Raspberry pi 4 falls too short on that...